requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv

sqlalchemy
psycopg2-binary

pandas
numpy
pyarrow

xarray
netCDF4

sentence-transformers
faiss-cpu

groq

plotly
streamlit

pytest
httpx
argopy
